Refactored application details
