+++ apps/partners-reference/src/applications/PaperApplicationDetails/DetailsApplicationContext.ts
@@ -1,0 +1,6 @@
+import React from "react"
+import { Application } from "@bloom-housing/core"
+
+export const DetailsApplicationContext = React.createContext<Application | null>(null)
+
+DetailsApplicationContext.displayName = "DetailsApplicationContext"
